Support displaying match dates
Add support for displaying match dates in the UI. This allows placing
trhe match date in a prominent position on the match page, where it is
easy for players and admins to see. Making the match date available in
the model will allow further features, such as sending notifications
before a match or reminders to submit scores afterwards.