Allow rescheduling matches
Add an interface to allow teams to request rescheduling matches, and for
the opposing team to approve or deny the request. This allows teams to
modify the match date without having to involve the admins. Match date
rescheduling is enabled on a per-league basis. It is disabled by default
for existing leagues and enabled by default for new leagues.

Rescheduling requests conceptually are split into two states: pending
and decided. Requests start pending, and cannot be edited. If a player
wishes to propose a new match date, existing requests from them are
self-denied. This prevents a user from requesting multiple match dates
at once (which could be considered to be a limitation). The design was
inspired by how transfers work.

Notifications are sent to captains when a request is created or denied,
and to all players when a request is approved. Admins can still edit the
match date directly, but this will not notify players. This should
reduce annoyance if admins want to fix historical match dates. If admins
want to change the match date and notify the players, they can create a
request and then accept it themselves. Alternatively, a 'send
notifications' checkbox on the match edit page would also work.

No validation of match dates is done. It is expected that wildly
unreasonable match dates will be rejected by the opposing team. Future
work could involve constraining match dates to the future, or to the
current week for "weekly" leagues.

Future work could involve integrating rescheduing with the existing
schedule system. For example, the site could suggest the next available
day, or flag unavailable days as such.